feature: add spatial mapping function
The spatial mapping takes a dataframe with just one column and distinct
spatial values. It then uses the spatial mappings stored in postgres
and makes matches for the spatial values to the corresponding spatialunit_uid

Currently the function is only designed to match for the levels country and canton.
The table it was developped for was a stattab px cube. So it should generally work
for all such px cubes that have no finer spatial granularity then the cantons.

The function was designed in such a way that it can easily be generalized to other spatial
mappings.