update spatial mapping and add new function for query testruns
the spatial mapping function has again an output that consists
only of the mapping to the `spatialunit_uid`.

A new function is added that allows to testrun queries against
an sqllite database. The database will only live in memeory
and it will consist of 2 tables: the table that the pipeline
creates and the spatial_unit table

The queries are expected to be in a `queries.sql` file in the
pipeline directory.

The output will be written in a `queries.log` file in the same
directory.

The queries in queries.sql should look like this:

```
-- SELECT COUNT(*) from -- Wieviele Abstimmungsvorlagen gab es seit dem Jahr 1971?
SELECT COUNT(*) as anzahl_vorlagen_in_2000
FROM abstimmungsvorlagen_seit_1971 as T
JOIN spatial_unit as S on T.spatialunit_uid = S.spatialunit_uid
WHERE S.name_de='Schweiz' AND S.country=TRUE
```

Empty lines will be ignored during parsing the queries.

The output in `queries.log` looks like this:

```
-- Wieviele Abstimmungsvorlagen gab es seit dem Jahr 1971?
SELECT COUNT(*) as anzahl_vorlagen_in_2000 FROM abstimmungsvorlagen_seit_1971 as T JOIN spatial_unit as S on
T.spatialunit_uid = S.spatialunit_uid WHERE S.name_de='Schweiz' AND S.country=TRUE;
=======================================================
  anzahl_vorlagen_in_2000
                    <int>
1                     624
```