adapt R fucntion code to postgres integration
- the postgres testruns are now performed on the postgres instance
once the dataset has been set to the status 'uploaded' otherwise
the testdb is set up in memory and the queries are run on the test db

- samples are only generated when the dataset is not in the status
'uploaded'

- metadata template are now generated by a seperate function into
template file: the real metadata files are not anymore overwriten
Next to metadata template generation there is a new function to read
the metadata

- the postgres integration has a function to connect to postgres and
anotherone to write the table for a dataset. The table is written
from the dataframe (tibble) of the dataset and derives the field
types from the metadata of the columns. Also a uid field is added
as the primary key. The foreign key to the spatial_unit can already
be defined during table creation

- the functions that load data from file have been moved inot their
own R script