R code: improve reading and discontinue get_path
the code to get the download url was hard coded to the German
language case for BFS. In all other cases it is clearer to specify
a `donwload_url` of `query_url` then to give only url parts in the
dataset list.

Also remove the `sheets` property from the dataset since this
property does not make sense anymore. In the code the property
was used to derive the `ds` class, the class intiation is now
done in a more understandable way.